+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.1.2] - 2023-09-22
+
 ### Changed
 
 - `useLoadDataState()` and `useLoadResourcesState()` to always pass an `AbortSignal`